requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
whitenoise==6.2.0



django-widget-tweaks==1.4.12
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-countries==7.5.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7

django-filter==22.1
django-forms-dynamic==1.0.0
django-formtools==2.3

django-htmx==1.13.0




django-rest-auth==0.9.5


django-widget-tweaks==1.4.12

